Provide an :over-ride adverb to the xml-element
This allows the over-riding of the name of the outer element of
a complex type that is being created for a class typed attribute.

If a name is provided to the trait that name will be used for the
element, otherwise the name will be derived from that of the attribute.